+++ app/src/main/java/com/duckduckgo/app/pixels/campaign/CampaignPixelParamsAdditionPlugin.kt
@@ -14,22 +14,20 @@
  * limitations under the License.
  */
 
-package com.duckduckgo.app.pixels.ppropromo
+package com.duckduckgo.app.pixels.campaign
 
-import com.duckduckgo.anvil.annotations.ContributesRemoteFeature
+import com.duckduckgo.anvil.annotations.ContributesPluginPoint
 import com.duckduckgo.di.scopes.AppScope
-import com.duckduckgo.feature.toggles.api.Toggle
 
-@ContributesRemoteFeature(
-    scope = AppScope::class,
-    featureName = "additionalPProPromoPixelParams",
-    settingsStore = AdditionalPproPixelParamFeatureStore::class,
-)
-interface AdditionalPproPixelParamsFeature {
+@ContributesPluginPoint(AppScope::class)
+interface CampaignPixelParamsAdditionPlugin {
+    /**
+     * @return extract the campaign name from the pixel query params
+     */
+    fun extractCampaign(queryParams: Map<String, String>): String?
 
     /**
-     * @return `true` to enable the privacy pro promo feature of appending a random subset of additional pixel parameters
+     * @return list of pixels that would be considered for pixel param addition
      */
-    @Toggle.DefaultValue(false)
-    fun self(): Toggle
+    fun names(): List<String>
 }
