+++ app/src/main/java/com/duckduckgo/app/pixels/campaign/params/AtpAdditionalPixelParamPlugin.kt
@@ -14,20 +14,19 @@
  * limitations under the License.
  */
 
-package com.duckduckgo.app.pixels.ppropromo
+package com.duckduckgo.app.pixels.campaign.params
 
 import com.duckduckgo.di.scopes.AppScope
+import com.duckduckgo.mobile.android.app.tracking.AppTrackingProtection
 import com.squareup.anvil.annotations.ContributesMultibinding
 import javax.inject.Inject
 
 @ContributesMultibinding(AppScope::class)
-class PproPixelParamsAdditionPlugin @Inject constructor(
-    private val additionalPixelParamsDataStore: AdditionalPixelParamsDataStore,
-) : PixelParamsAdditionPlugin {
-    override fun isEligible(queryParams: Map<String, String>): Boolean {
-        val origin = queryParams["origin"]
-        return origin != null && additionalPixelParamsDataStore.includedOrigins.contains(origin)
-    }
-
-    override fun names(): List<String> = listOf("m_subscribe")
+class AtpEnabledAdditionalPixelParamPlugin @Inject constructor(
+    private val appTrackingProtection: AppTrackingProtection,
+) : AdditionalPixelParamPlugin {
+    override suspend fun params(): Pair<String, String> = Pair(
+        "atpOnboarded",
+        "${appTrackingProtection.isOnboarded()}",
+    )
 }
