+++ app/src/main/java/com/duckduckgo/app/pixels/campaign/params/AutofillAdditionalPixelParams.kt
@@ -14,19 +14,24 @@
  * limitations under the License.
  */
 
-package com.duckduckgo.app.pixels.ppropromo.params
+package com.duckduckgo.app.pixels.campaign.params
 
-import com.duckduckgo.app.statistics.store.StatisticsDataStore
+import com.duckduckgo.autofill.impl.store.InternalAutofillStore
 import com.duckduckgo.di.scopes.AppScope
 import com.squareup.anvil.annotations.ContributesMultibinding
 import javax.inject.Inject
+import kotlinx.coroutines.flow.firstOrNull
 
 @ContributesMultibinding(AppScope::class)
-class ReinstallAdditionalPixelParamPlugin @Inject constructor(
-    private val statisticsDataStore: StatisticsDataStore,
+class AutofillUserAdditionalPixelParamPlugin @Inject constructor(
+    private val internalAutofillStore: InternalAutofillStore,
 ) : AdditionalPixelParamPlugin {
     override suspend fun params(): Pair<String, String> = Pair(
-        "isReinstall",
-        "${statisticsDataStore.variant == "ru"}",
+        "autofillUser",
+        "${getSavedPasswordCount() > 5}",
     )
+
+    private suspend fun getSavedPasswordCount(): Int {
+        return internalAutofillStore.getCredentialCount().firstOrNull() ?: 0
+    }
 }
